+++ src/main/java/com/salma/mini_projet_pharmacie/repository/PharmacienRepository.java
@@ -3,5 +3,5 @@ package com.salma.mini_projet_pharmacie.repository;
 import com.salma.mini_projet_pharmacie.model.Pharmacien;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PharmacienRepository extends JpaRepository<Pharmacien, Long> {
+public interface PharmacienRepository extends JpaRepository<Pharmacien, Integer> {
 }
