+++ src/main/java/com/salma/mini_projet_pharmacie/repository/ResponsableRepository.java
@@ -3,5 +3,5 @@ package com.salma.mini_projet_pharmacie.repository;
 import com.salma.mini_projet_pharmacie.model.Responsable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResponsableRepository extends JpaRepository<Responsable, Long> {
+public interface ResponsableRepository extends JpaRepository<Responsable, Integer> {
 }
